refactor(fe): separate katex rendering function
